Add files via upload
This assignment is a simple landing page of a fictional cinema named Kabul Ziba Cinema. It is my exercise on building a page layout using HTML and CSS and working on how the page will appear good across various screens. It has its navigation within the header, a welcome message on the hero, a section displaying what is on show (movies cards), location and lastly, the footer. I employed Flexbox to set up items and media queries to make the design responsive (work on the phone as well). The use of the color theme is dark with red accents in order to fit a movie style